% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagbs.R
\name{diag.bs}
\alias{diag.bs}
\title{Diagnostic Analysis - Local Influnce}
\usage{
envelope(model,k=19,alpha=0.05,res="deviance", precision = "fixed")
}
\arguments{
\item{model}{object of class \code{gamlss} holding the fitted model.}

\item{k}{number of replications for envelope construction. Default is 19.}

\item{alpha}{value giving the size of the envelope. Default is 0.05 which is equivalent to a 95\% band.}

\item{res}{type of residuals to be extracted. Default is deviance.}

\item{precision}{If \code{precision = "fixed"} a model with precision fixed is used;
else a model with precision variable is used.}
}
\value{
A simulated envelope of the class RBS.
}
\description{
Diagnostics for the RBS model
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
\references{
Atkinson, A. C. (1985) Plots, transformations and regression : an introduction to graphical methods of diagnostic regression analysis. Oxford Science Publications, Oxford.
}

