% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/envelopes.R
\name{envelope}
\alias{envelope}
\title{Simulated Evelope}
\usage{
envelope(model, k = 19, alfa = 0.05, res = "deviation")
}
\arguments{
\item{model}{object of class \code{gamlss} holding the fitted model.}

\item{k}{number of replications for envelope construction. Default is 19.}

\item{alfa}{value giving the size of the envelope. Default is 0.05 which is equivalent to a 95\% band.}

\item{res}{type of residuals to be extracted. Default is deviance.}
}
\description{
A normal plot with simulated envelope of the residual is produced.
}

