% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{Residuals}
\usage{
residuals(model,residual = "deviation")
}
\arguments{
\item{model}{an object for which the extraction of model residuals is meaningful.}

\item{residual}{type of residual to be used.}
}
\value{
Residuals extracted from the object object.
}
\description{
\code{residuals} is a function which extracts model residuals from objects returned by modeling functions.
}

