% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tests.R
\name{wald.test}
\alias{wald.test}
\title{Wald's test}
\usage{
lr.test(modelh0,modelh1)
}
\arguments{
\item{modelh1}{model under alternative hypothesis.}
}
\value{
A list with class "htest" containing the following components:

\code{statistic}	the value of the W-statistic.

\code{parameter}	the degrees of freedom for the W-statistic.

\code{p.value}	the p-value for the test.

\code{method}	a character string indicating what type of Wald test was performed.

\code{data.name} a character string giving the name(s) of the data
}
\description{
Performs a Wald test on vector of data.
}

