% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagbs.R
\name{diag.bs}
\alias{diag.bs}
\title{Diagnostic Analysis - Local Influnce}
\usage{
diag.bs(model)
}
\arguments{
\item{model}{object of class \code{gamlss} holding the fitted model.}
}
\value{
Local influence measures.
}
\description{
Diagnostics for the RBS model
}
\examples{
library(faraway)
data(cpd)
attach(cpd)
fit = gamlss(actual ~ 0+projected, family=RBS(mu.link="identity"),method=CG())
summary(fit)
set.seed(2015)
envelope(fit)
Cib <- diag.bs(fit)$Ci.beta
plot(Cib,ylim=c(0,1),pch=19)
abline(h=2*mean(Cib),type=2)
Cia <- diag.bs(fit)$Ci.alpha
plot(Cia,ylim=c(0,1),pch=19)
abline(h=2*mean(Cia),type=2)
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
\references{
Atkinson, A. C. (1985) Plots, transformations and regression : an introduction to graphical methods of diagnostic regression analysis. Oxford Science Publications, Oxford.
}

